+++ esl_proxy/include/algorithm/executor.h
@@ -22,9 +22,9 @@
  */
 typedef struct executor {
     uint8_t idx;
-    uint16_t tasks[AIC_OSTD];
-    uint16_t block_idx[AIC_OSTD];
-    uint16_t duration[AIC_OSTD];
+    uint32_t tasks[AIC_OSTD];
+    uint32_t block_idx[AIC_OSTD];
+    uint32_t duration[AIC_OSTD];
     uint64_t base[AIC_OSTD];
 } executor_t;
