+++ esl_proxy/src/algorithm/cutter.c
@@ -21,24 +21,24 @@ extern ctrl_t g_ctrl_t[DISPATCH_THREAD_CNT];
 extern _Atomic bool g_orch_is_done;
 extern _Atomic bool g_is_done;
 
-uint16_t  g_predecessor_cnt[RING_SIZE];
-uint16_t g_commit_task_id = 0;
-uint16_t g_completed_task_cnt = 0;
+uint32_t  g_predecessor_cnt[RING_SIZE];
+uint32_t g_commit_task_id = 0;
+uint32_t g_completed_task_cnt = 0;
 
-static inline bool update_task_state(uint16_t cnt, uint16_t* cq_buf)
+static inline bool update_task_state(uint32_t cnt, uint32_t* cq_buf)
 {
     if (cnt <= 0)
         return false;
     
-    uint16_t task_id;
-    uint16_t idx;
+    uint32_t task_id;
+    uint32_t idx;
     for (uint32_t j = 0; j < cnt; j++) {
         task_id = cq_buf[j];
         int idx = task_id;
         g_state_buf[idx].state = TASK_STATUS_COMPLETED;
     }
-    uint16_t i = atomic_load_explicit(&g_min_uncomplete_task, memory_order_acquire);
-    uint16_t end = atomic_load_explicit(&g_task_id, memory_order_acquire);
+    uint32_t i = atomic_load_explicit(&g_min_uncomplete_task, memory_order_acquire);
+    uint32_t end = atomic_load_explicit(&g_task_id, memory_order_acquire);
     for (; i < end; i++) {
         if (g_state_buf[i].state != TASK_STATUS_COMPLETED) {
             break;
@@ -49,13 +49,13 @@ static inline bool update_task_state(uint16_t cnt, uint16_t* cq_buf)
         g_min_uncomplete_task, end, g_ctrl_t[0].ready_queue[2].cnt, g_ctrl_t[0].ready_queue[1].cnt);
 }
 
-void add_successors(uint16_t ready_cnt[], uint16_t rq_buf[][RQ_BATCH_SIZE]) {
-    uint16_t end = atomic_load(&g_task_id);
-    uint16_t tmp = g_commit_task_id + PRE_BATCH_SIZE;
+void add_successors(uint32_t ready_cnt[], uint32_t rq_buf[][RQ_BATCH_SIZE]) {
+    uint32_t end = atomic_load(&g_task_id);
+    uint32_t tmp = g_commit_task_id + PRE_BATCH_SIZE;
     end = tmp > end ? end : tmp;
     while ( g_commit_task_id <= end)
     {
-        uint16_t task_idx = g_commit_task_id;
+        uint32_t task_idx = g_commit_task_id;
         struct predecessor_list *ptr = &g_predecessors[task_idx];
         if (ptr->cnt <= 0) {
             // WORKER_LOGF("ready, task_id,%u, task_idx,%u, ready_cnt,%u", g_commit_task_id, task_idx, *ready_cnt);
@@ -65,14 +65,14 @@ void add_successors(uint16_t ready_cnt[], uint16_t rq_buf[][RQ_BATCH_SIZE]) {
             WORKER_LOGF("ready_cnt[%d],%d",type, ready_cnt[type]);
             continue;
         }
-        uint16_t precessor_id = 0;
-        uint16_t predecessor_cnt = 0;
+        uint32_t precessor_id = 0;
+        uint32_t predecessor_cnt = 0;
         while (ptr->cnt > 0)
         {
             precessor_id = *(ptr->exp);
-            uint16_t precessor_idx = precessor_id;
+            uint32_t precessor_idx = precessor_id;
             if(g_state_buf[precessor_idx].state != TASK_STATUS_COMPLETED) {
-                uint16_t successor_idx = g_successor_buf[precessor_idx].cnt++;
+                uint32_t successor_idx = g_successor_buf[precessor_idx].cnt++;
                 g_successor_buf[precessor_idx].node[successor_idx] = g_commit_task_id;
                 g_state_buf[precessor_idx].successor_cnt++;
                 predecessor_cnt++;
@@ -93,8 +93,8 @@ void add_successors(uint16_t ready_cnt[], uint16_t rq_buf[][RQ_BATCH_SIZE]) {
     }
 }
 
-void send_2_ready_queue(uint16_t ready_cnt[], uint16_t rq_buf[][RQ_BATCH_SIZE]) {
-    for (uint16_t j = 0; j < 2; j++) {
+void send_2_ready_queue(uint32_t ready_cnt[], uint32_t rq_buf[][RQ_BATCH_SIZE]) {
+    for (uint32_t j = 0; j < 2; j++) {
         int target_ctrl = 0;
         queue_t *rq = &g_ctrl_t[target_ctrl].ready_queue[j];
         if (ready_cnt[j] > 0)
@@ -105,18 +105,18 @@ void send_2_ready_queue(uint16_t ready_cnt[], uint16_t rq_buf[][RQ_BATCH_SIZE]) 
     }
 }
 
-void resolve_dep(uint16_t cnt, uint16_t* cq_buf, uint16_t rq_buf[][RQ_BATCH_SIZE], uint16_t* ready_cnt) {
-    uint16_t task_id;
-    uint16_t succ_id;
-    uint16_t idx;
-    uint16_t succ_cnt;
+void resolve_dep(uint32_t cnt, uint32_t* cq_buf, uint32_t rq_buf[][RQ_BATCH_SIZE], uint32_t* ready_cnt) {
+    uint32_t task_id;
+    uint32_t succ_id;
+    uint32_t idx;
+    uint32_t succ_cnt;
     for (uint32_t j = 0; j < cnt; j++) {
         task_id = cq_buf[j];
         idx = task_id & RING_MASK;
         task_state st = g_state_buf[idx];
-        succ_cnt = (uint16_t)st.successor_cnt;
+        succ_cnt = (uint32_t)st.successor_cnt;
         WORKER_LOGF("completed,task_id,%u,type,%u, successor_cnt,%u", task_id, g_basic_buf[idx].type, succ_cnt);
-        for (uint16_t k = 0; k < succ_cnt; k++) {
+        for (uint32_t k = 0; k < succ_cnt; k++) {
             succ_id = g_successor_buf[idx].node[k];
             g_predecessor_cnt[succ_id & RING_MASK]--;
             WORKER_LOGF("cutter, task_id,%u, successor_id,%u, predecessor_cnt,%u", task_id, succ_id, g_predecessor_cnt[succ_id & RING_MASK]);
@@ -132,11 +132,11 @@ void resolve_dep(uint16_t cnt, uint16_t* cq_buf, uint16_t rq_buf[][RQ_BATCH_SIZE
 
 void deal_completed_queue() {
     for (int i = 0; i < DISPATCH_THREAD_CNT; i++) {
-        uint16_t cq_buf[CQ_BATCH_SIZE];
-        uint16_t rq_buf[2][RQ_BATCH_SIZE];
-        uint16_t ready_cnt[2] = {0, 0};
+        uint32_t cq_buf[CQ_BATCH_SIZE];
+        uint32_t rq_buf[2][RQ_BATCH_SIZE];
+        uint32_t ready_cnt[2] = {0, 0};
         queue_t *cq = &g_ctrl_t[i].completed_queue;
-        uint16_t cnt = CQ_BATCH_SIZE;
+        uint32_t cnt = CQ_BATCH_SIZE;
         batch_dequeue(cq, cq_buf, &cnt);
         g_completed_task_cnt += cnt;
         // for (size_t i = 0; i < cnt; i++)
