+++ esl_proxy/src/algorithm/dispatch.c
@@ -45,8 +45,8 @@ static inline void get_free_exe(int tid)
     set_mix(tid);
 }
 
-static inline void get_completed(uint64_t* bitmap, uint16_t task_id[], int *complete_cnt,
-                                 const uint16_t task_id_map[])
+static inline void get_completed(uint64_t* bitmap, uint32_t task_id[], int *complete_cnt,
+                                 const uint32_t task_id_map[])
 {
     int cnt = __builtin_popcountll(*bitmap);
     while (cnt > 0) {
@@ -62,7 +62,7 @@ static inline void get_completed(uint64_t* bitmap, uint16_t task_id[], int *comp
 // TODO: add counter for spmd
 static inline void push_2_completed_queue(int tid)
 {
-    uint16_t task_id[240];
+    uint32_t task_id[240];
     int complete_cnt = 0;
     for (int i = 0; i < EXE_TYPE_CNT; i++) {
         get_completed(&g_ctrl_t[tid].msg_bitmap[i][0], task_id, &complete_cnt,
@@ -70,7 +70,7 @@ static inline void push_2_completed_queue(int tid)
         get_completed(&g_ctrl_t[tid].msg_bitmap[i][1], task_id, &complete_cnt,
                       g_ctrl_t[tid].task_id_map2[i]);
     }
-    batch_enqueue(&g_ctrl_t[tid].completed_queue, task_id, (uint16_t)complete_cnt);
+    batch_enqueue(&g_ctrl_t[tid].completed_queue, task_id, (uint32_t)complete_cnt);
     atomic_fetch_add_explicit(&g_completed_cnt, complete_cnt, memory_order_acquire);
 }
 
@@ -85,14 +85,14 @@ static inline int send_task(ctrl_t *ctrl, int type)
         WORKER_LOGF("send,free_cnt,%d", cnt);
         return 0;
     }
-    uint16_t task_ids[AIC_CNT];
+    uint32_t task_ids[AIC_CNT];
     if (!batch_dequeue(&ctrl->ready_queue[type], task_ids, &cnt)){
         return 0;
     }
     
     int sent = 0;
     for (int i = 0; i < cnt; i++) {
-        uint16_t task_id = task_ids[i];
+        uint32_t task_id = task_ids[i];
         uint64_t idx = (uint64_t)__builtin_ctzll(free_bitmap);
 
         uint64_t mask = (uint64_t)0x1 << idx;
